Tweaks for Adafruit_GFX

+++ examples/ssd1306/ssd1306.pde
@@ -182,7 +182,7 @@ void testdrawchar(void) {
 }
 
 void testdrawcircle(void) {
-  for (uint8_t i=0; i<display.height(); i+=2) {
+  for (int16_t i=0; i<display.height(); i+=2) {
     display.drawCircle(display.width()/2, display.height()/2, i, WHITE);
     display.display();
   }
@@ -190,7 +190,7 @@ void testdrawcircle(void) {
 
 void testfillrect(void) {
   uint8_t color = 1;
-  for (uint8_t i=0; i<display.height()/2; i+=3) {
+  for (int16_t i=0; i<display.height()/2; i+=3) {
     // alternate colors
     display.fillRect(i, i, display.width()-i*2, display.height()-i*2, color%2);
     display.display();
@@ -199,7 +199,7 @@ void testfillrect(void) {
 }
 
 void testdrawtriangle(void) {
-  for (uint16_t i=0; i<min(display.width(),display.height())/2; i+=5) {
+  for (int16_t i=0; i<min(display.width(),display.height())/2; i+=5) {
     display.drawTriangle(display.width()/2, display.height()/2-i,
                      display.width()/2-i, display.height()/2+i,
                      display.width()/2+i, display.height()/2+i, WHITE);
@@ -220,7 +220,7 @@ void testfilltriangle(void) {
 }
 
 void testdrawroundrect(void) {
-  for (uint8_t i=0; i<display.height()/2-2; i+=2) {
+  for (int16_t i=0; i<display.height()/2-2; i+=2) {
     display.drawRoundRect(i, i, display.width()-2*i, display.height()-2*i, display.height()/4, WHITE);
     display.display();
   }
@@ -228,7 +228,7 @@ void testdrawroundrect(void) {
 
 void testfillroundrect(void) {
   uint8_t color = WHITE;
-  for (uint8_t i=0; i<display.height()/2-2; i+=2) {
+  for (int16_t i=0; i<display.height()/2-2; i+=2) {
     display.fillRoundRect(i, i, display.width()-2*i, display.height()-2*i, display.height()/4, color);
     if (color == WHITE) color = BLACK;
     else color = WHITE;
@@ -237,51 +237,51 @@ void testfillroundrect(void) {
 }
    
 void testdrawrect(void) {
-  for (uint8_t i=0; i<display.height()/2; i+=2) {
+  for (int16_t i=0; i<display.height()/2; i+=2) {
     display.drawRect(i, i, display.width()-2*i, display.height()-2*i, WHITE);
     display.display();
   }
 }
 
 void testdrawline() {  
-  for (uint8_t i=0; i<display.width(); i+=4) {
+  for (int16_t i=0; i<display.width(); i+=4) {
     display.drawLine(0, 0, i, display.height()-1, WHITE);
     display.display();
   }
-  for (uint8_t i=0; i<display.height(); i+=4) {
+  for (int16_t i=0; i<display.height(); i+=4) {
     display.drawLine(0, 0, display.width()-1, i, WHITE);
     display.display();
   }
   delay(250);
   
   display.clearDisplay();
-  for (uint8_t i=0; i<display.width(); i+=4) {
+  for (int16_t i=0; i<display.width(); i+=4) {
     display.drawLine(0, display.height()-1, i, 0, WHITE);
     display.display();
   }
-  for (int8_t i=display.height()-1; i>=0; i-=4) {
+  for (int16_t i=display.height()-1; i>=0; i-=4) {
     display.drawLine(0, display.height()-1, display.width()-1, i, WHITE);
     display.display();
   }
   delay(250);
   
   display.clearDisplay();
-  for (int8_t i=display.width()-1; i>=0; i-=4) {
+  for (int16_t i=display.width()-1; i>=0; i-=4) {
     display.drawLine(display.width()-1, display.height()-1, i, 0, WHITE);
     display.display();
   }
-  for (int8_t i=display.height()-1; i>=0; i-=4) {
+  for (int16_t i=display.height()-1; i>=0; i-=4) {
     display.drawLine(display.width()-1, display.height()-1, 0, i, WHITE);
     display.display();
   }
   delay(250);
 
   display.clearDisplay();
-  for (uint8_t i=0; i<display.height(); i+=4) {
+  for (int16_t i=0; i<display.height(); i+=4) {
     display.drawLine(display.width()-1, 0, 0, i, WHITE);
     display.display();
   }
-  for (uint8_t i=0; i<display.width(); i+=4) {
+  for (int16_t i=0; i<display.width(); i+=4) {
     display.drawLine(display.width()-1, 0, i, display.height()-1, WHITE); 
     display.display();
   }
